whisper-worker/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

openai-whisper>=20231117
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<0.28.0
yt-dlp>=2024.7.2
python-multipart>=0.0.9
